Refactor connection to DB + Validating the whole project

+++ InventorySystem/InventorySystem/Products.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Configuration;
 using System.Data;
 using System.Data.SqlClient;
 using System.Drawing;
@@ -13,6 +14,7 @@ namespace InventorySystem
 {
     public partial class Products : Form
     {
+
         public Products()
         {
             InitializeComponent();
@@ -26,38 +28,41 @@ namespace InventorySystem
 
         private void button2_Click(object sender, EventArgs e)
         {
-            //TODO Check if valid login
-            SqlConnection con = new SqlConnection("Data Source=.;Initial Catalog=InventorySystem;Integrated Security=True");
-            // Insert Logic
-            con.Open();
-            bool status = false;
-            if (comboBox1.SelectedIndex == 0)
+            if (Validation())
             {
-                status = true;
-            }
-            else
-            {
-                status = false;
-            }
+                SqlConnection con = Connection.GetConnection("WebDev");
+                // Insert Logic
+                con.Open();
+                bool status = false;
+                if (comboBox1.SelectedIndex == 0)
+                {
+                    status = true;
+                }
+                else
+                {
+                    status = false;
+                }
 
-            var sqlQuery = "";
-            if (IfProductExists(con, textBox1.Text))
-            {
-                sqlQuery = @"UPDATE [Products] SET [ProductName] = '" + textBox2.Text + "',[ProductStatus] = '" + status +
-                          "' WHERE [ProductCode] = '" + textBox1.Text + "'";
-            }
-            else
-            {
-                sqlQuery = @"INSERT INTO [dbo].[Products] ([ProductCode] ,[ProductName] ,[ProductStatus])
+                var sqlQuery = "";
+                if (IfProductExists(con, textBox1.Text))
+                {
+                    sqlQuery = @"UPDATE [Products] SET [ProductName] = '" + textBox2.Text + "',[ProductStatus] = '" + status +
+                              "' WHERE [ProductCode] = '" + textBox1.Text + "'";
+                }
+                else
+                {
+                    sqlQuery = @"INSERT INTO [dbo].[Products] ([ProductCode] ,[ProductName] ,[ProductStatus])
                              VALUES ('" + textBox1.Text + "','" + textBox2.Text + "','" + status + "');";
+                }
+
+                SqlCommand cmd = new SqlCommand(sqlQuery, con);
+                cmd.ExecuteNonQuery();
+                con.Close();
+
+                //Reading Data
+                LoadData();
+                ResetRecords();
             }
-
-            SqlCommand cmd = new SqlCommand(sqlQuery, con);
-            cmd.ExecuteNonQuery();
-            con.Close();
-
-            //Reading Data
-            LoadData();
         }
 
         private bool IfProductExists(SqlConnection Connection, string ProductCode)
@@ -65,66 +70,54 @@ namespace InventorySystem
             SqlDataAdapter sda = new SqlDataAdapter("select 1 from [Products] WHERE [ProductCode]='" + ProductCode + "'", Connection);
             DataTable dt = new DataTable();
             sda.Fill(dt);
-            if (dt.Rows.Count > 0)
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
+            return dt.Rows.Count > 0;
         }
 
         public void LoadData()
         {
-            SqlConnection con = new SqlConnection("Data Source=.;Initial Catalog=InventorySystem;Integrated Security=True");
-            SqlDataAdapter sda = new SqlDataAdapter("select * from [dbo].[Products]", con);
+            SqlConnection con = Connection.GetConnection("WebDev");
             DataTable dt = new DataTable();
-            sda.Fill(dt);
-            dataGridView1.Rows.Clear();
-            foreach (DataRow items in dt.Rows)
-            {
-                int n = dataGridView1.Rows.Add();
-                dataGridView1.Rows[n].Cells[0].Value = items["ProductCode"].ToString();
-                dataGridView1.Rows[n].Cells[1].Value = items["ProductName"].ToString();
-                if ((bool)items["ProductStatus"])
-                {
-                    dataGridView1.Rows[n].Cells[2].Value = "Active";
-                }
-                else
-                {
-                    dataGridView1.Rows[n].Cells[2].Value = "Inactive";
-                }
-            }
+            dt = DA_adapter.GetTable();
+            BindingSource bs = new BindingSource(dt, null);
+            dataGridView1.DataSource = dt;
         }
 
         private void button1_Click(object sender, EventArgs e)
         {
-            SqlConnection con = new SqlConnection("Data Source=.;Initial Catalog=InventorySystem;Integrated Security=True");
-            var sqlQuery = "";
-            if (IfProductExists(con, textBox1.Text))
+            DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete this column ?", "Message", MessageBoxButtons.YesNo);
+            if (DialogResult == DialogResult.Yes)
             {
-                con.Open();
-                sqlQuery = @"DELETE FROM [Products] WHERE [ProductCode] = '" + textBox1.Text + "'";
+                if (Validation())
+                {
+                    SqlConnection con = Connection.GetConnection("WebDev");
+                    var sqlQuery = "";
+                    if (IfProductExists(con, textBox1.Text))
+                    {
+                        con.Open();
+                        sqlQuery = @"DELETE FROM [Products] WHERE [ProductCode] = '" + textBox1.Text + "'";
 
-                SqlCommand cmd = new SqlCommand(sqlQuery, con);
-                cmd.ExecuteNonQuery();
-                con.Close();
-            }
-            else
-            {
-                MessageBox.Show("Record Not Exist...!!");
-            }
+                        SqlCommand cmd = new SqlCommand(sqlQuery, con);
+                        cmd.ExecuteNonQuery();
+                        con.Close();
+                    }
+                    else
+                    {
+                        MessageBox.Show("Record Not Exist...!!");
+                    }
 
-            //Reading Data
-            LoadData();
+                    //Reading Data
+                    LoadData();
+                    ResetRecords();
+                }
+            }
         }
 
         private void dataGridView1_MouseClick(object sender, MouseEventArgs e)
         {
+            button2.Text = "Update";
             textBox1.Text = dataGridView1.SelectedRows[0].Cells[0].Value.ToString();
             textBox2.Text = dataGridView1.SelectedRows[0].Cells[1].Value.ToString();
-            if (dataGridView1.SelectedRows[0].Cells[2].Value.ToString() == "Active")
+            if (dataGridView1.SelectedRows[0].Cells[2].Value.ToString() == "True")
             {
                 comboBox1.SelectedIndex = 0;
             }
@@ -132,6 +125,45 @@ namespace InventorySystem
             {
                 comboBox1.SelectedIndex = 1;
             }
+        }
+
+        private void ResetRecords()
+        {
+            textBox1.Clear();
+            textBox2.Clear();
+            comboBox1.SelectedIndex = -1;
+            button2.Text = "Add";
+            textBox1.Focus();
+        }
+
+        private void button3_Click(object sender, EventArgs e)
+        {
+            ResetRecords();
+        }
+
+        private bool Validation()
+        {
+            bool result = false;
+            if (string.IsNullOrEmpty(textBox1.Text))
+            {
+                errorProvider1.Clear();
+                errorProvider1.SetError(textBox1, "Product Code Required !");
+            }
+            else if (string.IsNullOrEmpty(textBox2.Text))
+            {
+                errorProvider1.Clear();
+                errorProvider1.SetError(textBox2, "Product Name Required !");
+            }
+            else if (comboBox1.SelectedIndex == -1)
+            {
+                errorProvider1.Clear();
+                errorProvider1.SetError(comboBox1, "Select Status !");
+            }
+            else
+            {
+                result = true;
+            }
+            return result;
         }
     }
 }
